Add SSE files. Rename a few things.

+++ Tuniac1/TuniacApp/SSE_Utils.h
@@ -1,6 +1,6 @@
 #pragma once
 
-void SSE_CopyFloat(float * Src, float * Dest, int len);
+void SSE_CopyFloat(float * Src, float * Dest, unsigned long len);
 void SSE_ClearFloat(float * Array, int len);
 
 void SSE_AddArrayStore(float * ArrayA, float * ArrayB, float * storage, int len);
